backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
websockets>=12.0

# Pydantic for validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP client for OpenRouter
httpx>=0.28.0

# Environment variables
python-dotenv>=1.0.1

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
